- profile

+++ app/src/main/java/com/alekseyld/watertraker/model/Drink.kt
@@ -4,4 +4,4 @@ import android.graphics.Bitmap
 
 data class Drink(val name: String,
                  val image: Bitmap?,
-                 val koef: Float)
+                 val koef: Double)
